web_docker/requirements.txt
flask==3.0.0
werkzeug==3.0.1
pillow>=10.0.0
imagehash>=4.3.1
ultralytics>=8.0.0
opencv-python-headless>=4.6.0
numpy>=1.23.0
torch>=2.0.0
torchvision>=0.15.0